Create issue-cfn-signal.yaml
To ensure that CloudFormation correctly fails when the signal is not received successfully, we need to introduce a WaitCondition and associate it with a WaitHandle. The EC2 instance will then send a success or failure signal to the WaitCondition, which CloudFormation will use to determine whether the stack deployment succeeds or fails.